Add get Total like, Fix Bug Playlist and Playlist Detail, Add duration video, Add 2 component in vuejs (edit-playlist and drag-items), fix bug template

+++ resources/js/components/subscribe-button.js
@@ -47,26 +47,22 @@ Vue.component('subscribe-button', {
                 return alert('Login To Subscribe')
             }
 
-            // if(this.owner) {
-            //     return alert('You cannot subscribe to your channel')
-            // }
+            if(this.owner) {
+                return alert('You cannot subscribe to your channel')
+            }
 
             if(this.subscribed) {
                 axios.delete(`/channels/${this.channel.id}/subscriptions/${this.subscription.id}`)
                     .then(() => {
-                        // console.log(this.subscriptions)
                         this.subscriptions = this.subscriptions.filter(s => s.id !== this.subscription.id)
-                        // console.log(this.subscriptions)
                     })
             } else {
                 axios.post(`/channels/${this.channel.id}/subscriptions`)
                     .then(response => {
-                        // console.log(response)
                         this.subscriptions = [
                             ...this.subscriptions,
                             response.data
                         ]
-                        // console.log(...this.subscriptions)
                     })
             }
         }
